Move all logging related function to specific files.
