+++ cpp/src/bindings.cpp
@@ -8,11 +8,9 @@
 #include <pybind11/stl.h>
 #include <pybind11/numpy.h>
 #include "pypowsybl.h"
+#include "pylogging.h"
 
 namespace py = pybind11;
-
-pypowsybl::CppToPythonLogger *pypowsybl::CppToPythonLogger::singleton_ = nullptr;
-bool pypowsybl::CppToPythonLogger::initialized_ = false;
 
 template<typename T>
 void bindArray(py::module_& m, const std::string& className) {
@@ -586,6 +584,6 @@ PYBIND11_MODULE(_pypowsybl, m) {
     m.def("get_three_windings_transformer_results", &pypowsybl::getThreeWindingsTransformerResults,
           "create a table with all three windings transformer results computed after security analysis", py::arg("result"));
     m.def("create_element", ::createElement, "create a new element on the network", py::arg("network"),  py::arg("dataframes"),  py::arg("elementType"));
-    m.def("set_logger", &pypowsybl::setLogger, "Setup the logger", py::arg("logger"));
-    m.def("get_logger", &pypowsybl::getLogger, "Retrieve the logger");
+    m.def("set_logger", &setLogger, "Setup the logger", py::arg("logger"));
+    m.def("get_logger", &getLogger, "Retrieve the logger");
 }
